+++ DatadogCore/Sources/Core/DatadogCore.swift
@@ -429,17 +429,10 @@ extension DatadogContextProvider {
         subscribe(\.launchTime, to: LaunchTimePublisher())
         #endif
 
-        if #available(iOS 12, tvOS 12, *) {
-            subscribe(\.networkConnectionInfo, to: NWPathMonitorPublisher())
-        } else {
-            assign(reader: SCNetworkReachabilityReader(), to: \.networkConnectionInfo)
-        }
+        subscribe(\.networkConnectionInfo, to: NWPathMonitorPublisher())
+
         #if os(iOS) && !targetEnvironment(macCatalyst) && !(swift(>=5.9) && os(visionOS))
-        if #available(iOS 12, *) {
-            subscribe(\.carrierInfo, to: iOS12CarrierInfoPublisher())
-        } else {
-            assign(reader: iOS11CarrierInfoReader(), to: \.carrierInfo)
-        }
+        subscribe(\.carrierInfo, to: CarrierInfoPublisher())
         #endif
 
         #if os(iOS) && !targetEnvironment(simulator)