+++ DatadogCore/Tests/Datadog/Mocks/SystemFrameworks/CoreTelephonyMocks.swift
@@ -53,20 +53,11 @@ class CTTelephonyNetworkInfoMock: CTTelephonyNetworkInfo {
             "000001": CTCarrierMock(carrierName: newCarrierName, isoCountryCode: newISOCountryCode, allowsVOIP: newAllowsVOIP)
         ]
 
-        if #available(iOS 12.0, *) {
-            serviceSubscriberCellularProvidersDidUpdateNotifier?("000001")
-        }
+        serviceSubscriberCellularProvidersDidUpdateNotifier?("000001")
     }
-
-    // MARK: - iOS 12+
 
     override var serviceCurrentRadioAccessTechnology: [String: String]? { _serviceCurrentRadioAccessTechnology }
     override var serviceSubscriberCellularProviders: [String: CTCarrier]? { _serviceSubscriberCellularProviders }
-
-    // MARK: - Prior to iOS 12
-
-    override var currentRadioAccessTechnology: String? { _serviceCurrentRadioAccessTechnology?.first?.value }
-    override var subscriberCellularProvider: CTCarrier? { _serviceSubscriberCellularProviders?.first?.value }
 }
 
 #endif